Move SearchBar para diretorio de components
